Replace Stripe with Paddle: frontend
Replace Stripe Elements + confirmSetup flow with Paddle
inline checkout opened via transactionId. Card details
now arrive asynchronously via webhook rather than
synchronously from the setup intent result.
